Revert "Add analytics feature flag"
We can foresee situations where we'd want to turn off Google Analytics,
even after the feature is stable. For this reason, we think it makes
sense to have it as a top-level setting, rather than a feature flag. We
already do something similar for the Cloudwatch metrics setting.

This commit reverts the feature flag - we'll implement the top-level
setting in the next commit.

This reverts commit 1771fcff6ed9810792b35fd6ab792c8e420ed989.