Add google analytics behind flag
Adds a subset of the analytics code from the product pages repo to this
one and runs it when the settings flag is enabled.

We'll need more of the code from the product pages later to deal with
the cookie consent, but that's being handled in a separate card so I'm
eeping it out of this PR for tidiness.