Move cookie banner into component
The cookie banner is an isolated enough piece of UI that it makes sense
to split it into its own component.

We can also move the cookie banner JS into the component folder since
it's quite coupled to this UI so it seems neater organisation-wise.

As part of this we're also adding translations for the text.