Add /security.txt to routes
This adds a /.well-known/security.txt in line with the gds way, which helps people report vulnerabilities to cyber.
It does this by redirecting to alphagov's security.txt from /security.txt and /.well-known/security.txt

This commit updates a previous mistake where I was using the endpoint /security.txt.html and /.well-known/security.txt.html. This commit removes the .html