Add custom logger class that attaches default context
Add an implementation of the
`ActiveSupport::CurrentAttributes` abstract super class,
`CurrentLoggingAttributes` for storing attributes that will be
attached to all log lines by default if they are present.

`ActiveSupport::CurrentAttributes` provides a thread-isolated
attributes singleton, which resets automatically before and after each
request.

Add a customer logger class - `ApplicationLogger` which extracts
attributes from the `CurrentLoggingAttributes` to add the stored
attributes to application log lines.

This will replace our current methodology of passing around a
`logging_context` object that we store on the `ApplicationController`
and add to during the processing of the request.

The intention is to simplify our logging approach, so develpoers do
not need to be concerned with passing around the `logging_context`
correctly. This will also mean that any logging we do using
`Rails.logger.info` etc will receive the context automatically,
allowing for us to do more ad-hoc logging without needing to use
the `EventLogger`, if we desire.