Allow title helper to take multiple arguments
On the generic pages (i.e. those which aren't part of a form flow),
we want the title to take the format `#{page} – GOV.UK Forms – GOV.UK`.

We could do this manually, but to make it as easy and consistent as
possible it makes sense to update the page title helpers to handle
multiple string arguments.