Add service to write submission to CSV
Adds SubmissionCsvService, which, given a filename and the session
context for the form writes the submission to a CSV file.

The first column is the form reference, followed by a column per
question where the header is the question name.

This uses the Ruby "csv" library, which offers basic CSV
reading/writing functionality.

The outputted CSV file will use the default unix line endings and
UTF-8 encoding.