Write submission as CSV if feature flag enabled
If the `attach_csv_to_submission_email` feature flag is enabled, write
the submission to a CSV in a temporary file.

In future, we will send the CSV to form processors using Notify.

We will also enhance the feature flag so it can be enabled on a form
by form basis later.