Allow FeatureService to switch features per form
This commit allows features to be toggled on a per form basis in
addition to the global feature flag. This is done in a similar way as
per org switching in forms-admin. You can override the default value of
a feature flag by providing key value pairs under a forms object, where
the key is the form ID. For example:

```
features:
  flag_test:
    enabled: false
    forms:
      "123": true
```

We'd like to use this to do some user research for additional
submission types with specific forms.