Attach CSV to submission email
Generate a CSV file with the submission data and save to a temporary
file if the feature flag is enabled. Upload this file to Notify when
we send the submission email. The link to the file is available in
the Notify template using the `link_to_file` personalisation field.

Also include a `csv_attached` boolean field in the personalisation so
that we can include optional content in the template when a CSV file
is attached.