Change all url helpers to use keyword arguments
Currently the feature tests call url helpers, like form_page_path, with
a mix of positional and keyword arguments .

An example of a positional argument call:
form_page_path("preview-draft", 2, form_response_data.form_slug, 1)

An example of a keyword call:
form_page_path(mode: "preview-draft", form_id: 2, form_slug: form_response_data.form_slug, page_slug: 1)

This is fine, but inconsistent.

This PR changes all calls to use the keyword form. This makes it clearer
which arguments are for which routing segment. The keyword argument is
clearer because there are 4 arguments which makes it hard to get right.

It also allows adding new optional parameters to the calls, which will
be useful in the future.

There maybe a better way to do this in general, rather than using the
helpers with all params like this.

This commit does the simplest thing to make it consistent.