Make Page factory work with ActiveResource
The Page factory is mainly used to generate JSON payloads when mocking
the API.

It's not currently used to test against directly. One reason for this is
that a Page built using FactoryBot behaves differently than one which
has been fetched from the API.

One difference is how they respond to attributes which don't exist.

For example:
```ruby
fb_page = build :page
fb_page.non_existent_attribute
```

will return `nil`.

Mocking the JSON response from the API and then trying to call
mocked_page.non_existent_attribute will throw a NoMethod exception.
This matches what would happen in the app code.

This difference is due to the way FactoryBot builds objects. By default,
FactoryBot will create a new object, then assign each attribute
specified in the factory to the new object.

This works for ActiveRecord but not ActiveResource.

It doesn't work for ActiveResource. When ActiveResource fetches an object
using an API call, it performs the API call to get the attributes then
passes them to the .new method.

FactoryBot [provides a way to override the way it builds objects](https://github.com/thoughtbot/factory_bot/blob/main/GETTING_STARTED.md#custom-construction).

This commit uses it to change the way factories for the Page model are
built.

This change means that the factory can be used when testing the model.

It should allow the factory to be used more widely across our tests.