Ensure Page factory is valid
The Page factory relies on constants which are defined in the API codebase, not in the
runner.

Calling build(:page) throws an error.

Because the factory is only ever called with a specific answer_type,
it doesn't raise an error in any of our current tests.

This commit removes the code, replacing it with "number" for now and
adds a test to ensure the factory is valid.