Change answer_settings tests to use factory
The tests for Page#answer_settings use mocks to check for valid results.

This commit changes them to use the factory instead.

We have to remove answer_settings from the factory because it's hard to
completely remove an attribute from a factory once it's included. In
this case that makes sense, because the code we are testing ensures that
calling answer_settings always returns a value.