Reset answer if question is changed to repeating
When a form creator is editing questions, they may change value of
is_repeatable.

For example, this could happen When previewing a form. Making a
repeatable question not repeatable would cause an error if the user
already has a value in the session.

When this happens, the value of the session will be an array. This
cannot be passed to update_attributes, which throws an argument error.

This commit fixes this case by catching this error when a users' answers
are loaded.

Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@digital.cabinet-office.gov.uk>
Co-authored-by: Samuel Culley <samuel.culley@digital.cabinet-office.gov.uk>