Fail tests when there are console errors
In some cases, when the tests write to stderror, we do not get a test
failure - instead vitest just accepts it and moves on to the next test.

There might be valid reasons for vitest doing this (e.g. it allows you
to test error output) but it's not a use case we want to support at the
moment.

To fix this, in a beforeEach block we spy on console.error and mock its
implementation to throw an error in the test context. This means it'll
throw a real error and fail the test, rather than just print to the
console.