Retry sending submission email without file upon error
When the NotificationClient raises a
`Notifications::Client::BadRequestError` upon attempting to deliver
the submission email when a CSV file is attached, try the request
again without the CSV file attached.

The intention of this is to allow submissions to go through if there
is an unexpected error with the CSV file attachment. This could be
due to the file failing a virus scan, or another unanticipated issue
with the file we have attached.

Send the exception to Sentry and log at error level when this happens,
as we want to be notified if we fail to attach the CSV file.