Raise action mailer delivery errors in development
It makes sense to raise delivery errors in the development enviornment
so it more closely reflects the production behaviour. A less
informative error is currently thrown if delivery fails due to, for
example, the Notify API key not being set.

This also allows us to locally test that we retry delivery without a
CSV file attached if delivery fails with a CSV file attached.

Also update `production.rb` to explicitly set
`config.action_mailer.raise_delivery_errors` to true. This is the
default because the line was commented out, but it makes sense to be
explicit and remove the misleading comment.