Add new title for add another page
The add another page currently has the same page title as the question
page.

This would fail an accessibility audit because it's not a unique title
which describes the page and its relationship to others. When a user
navigates from the question to the add another page, the title wouldn't
change.

This is the most common user journey and we should give users
context to let them know they are on a new page. This is
especially important for form fillers using screen readers as they are
more likely to use page titles when navigating.

This commit changes the page title for the add another screen.

It's normally good practice to make the page title match the page
heading. The heading in this case is "You have added one answer".

The heading doesn't provide much context for the user to know where they
are. This may change in the future.

This commit is just to fix the current issue with page titles.