Change max number answers for repeatable_step
The current limit of ten answers for a single repeatable question is
being increased from 10 to 50.

This figure has been chosen to cover most use cases for form creators.

This commit changes the MAX_ANSWERS constant from 10 to 50 and fixes the
tests which also need to change.

The tests could be changed to not break when changing this constant.

We are unlikely to change this number often. The tests have been kept as
explicitly checking the text and value.