Add service object to convert from v2 to v1 forms
The safest and quickest way to get this app using v2 form documents is
to start by converting them to v1 form snapshots and then handing those
over to the existing code.

This commit adds a service object that does this conversion, similar to
the approach taken in alphagov/forms-api#563.