Add object for retrieving v2 form documents
Add `Api::V2::FormDocumentResource`. This an ActiveResource model that
can be used with the v2 API described in [ADR034].

This commit adds the model in a new folder, `/app/resources`, rather
than `/app/models`, because we don't want to use it as a model currently
(we're going to convert the results to a v1 form immediately), and we
may want to decouple the in memory representation of forms from the wire
representation in future.

[ADR034]: https://github.com/alphagov/forms/blob/main/ADR/ADR034-api-v2-for-forms.md