Allow using v2 API
Change `Api::V1::FormSnapshotRepository` to use
`Api::V2::FormDocumentResource` and `Api::V1::Converter` to allow this
app to retrieve forms using the v2 API without changes to the rest of
the codebase. Basically, we get the v2 form documents and then
immediately convert them to the older form snapshot format so we can
keep using the same models.

This behaviour is currently behind a feature flag and not enabled.