Add service object for getting v1 form snapshots
This commit adds an extra layer of abstraction around how we retrieve
forms from the forms-api app.

The thinking behind this is that it will make swapping out the method
used for getting those forms easier. It also lets us in future decouple
our models from the way we get the data, which is different to how
ActiveResource does things but might be preferable.