Add feature flag for using v2 API
It would be good to be able to switch this app over to using the new v2
API in a test environment before doing so in production, and also to be
able to quickly revert the change if necessary.

This commit adds a feature flag to this end. Hopefully it won't stick
around for long.