Reinstate simple feature flagging
This commit removes per form feature flagging. This was originally
intended to allow a gradual rollout of additional submission types, but
we found that storing the submission_type on the form itself was
simpler, and is a longer term solution. We recommend that future per
form flagging is done via the database.

Co-authored-by: Stephen Daly <stephen.daly@digital.cabinet-office.gov.uk>