Support repeated steps in submission CSV
When there is a repeated step in the completed form - i.e. when a
question that allows multiple answers has been answered multiple
times, include a column in the submission CSV for each repeated
answer. The column header is suffixed by the answer index, like
' - Answer 1'.

For example,

'What is the meaning of life - Answer 1, What is the meaning of life -
 Answer 2'

When the question has separate attributes for different parts of the
answer, like a name type question can, the suffix comes after the
suffix for the attribute name.

For example,

'What is your name - First name - Answer 1, What is your name - Last
name - Answer 1'