Remove api_v2 feature flag
Because this was the only feature flag in the code at the moment I've
had to change the settings specs to test that the features settings has
an empty default value. This is how it was last time we had no feature
flags present (see commit 16aff23a5). If we add some more flags we can
change the tests back again.