Add autoselect config
We'll need to set the autocomplete's `autoselect` config setting to
false when we implement long lists. This is because we can't know enough
about each list's options to automatically choose the first one on a
form filler's behalf.

To implement this, this commit adds autoselect as a configurable option
in the autocomplete helper.