Add tests for Step model
The Step model doesn't have any unit tests.

This makes changing the class harder:
- it's tricky to know we haven't broken something the code relies on
- new methods are less likely to have tests written for them if there is
  no exsiting spec file to add to

This commit adds a spec file for the Step model.

The tests are writted to match the existing code and the functionality
it currently has.

There are ways we could improve the Step model in the future which would
make the tests easier to write and maintain.

These tests will be the base on which improvements can be made.

For example, we could:
- make some methods in the Step model private
- remove any methods which are not used in the code
- improve the way the Step model is instantiated to make it easier to
  configure and test
- add a model for conditions, which is currently an implicit class
  created by ActiveResource
- clear up the factories to make them easier to use in tests, rather
  than using doubles