Specify S3 bucket region
When using the AWS SDK for Ruby, it is required to provide the bucket
region. If the region is different to the region the bucket was
created in, we get the following error:

Aws::s3::Errors::PermanentRedirect - The bucket you are attempting to
access must be addressed using the specified endpoint. Please send all
future requests to this endpoint.

We now store the s3_bucket_region on the form. Use this when
initializing the S3 client so that we make requests to S3 using the
region for the bucket.