Extract services for different submission types
To simplify FormSubmissionService, extract the implementation of
sending the submission to the processing team for the different
submission types into separate services.

Everything to do with sending the submission using Notify is now in
NotifySubmissionService, and for S3 it is in S3SubmissionService.

This also ensures that we log submission when:
- The submission type is S3
- When a preview form is submitted and the submission email isn't set.
Technically the form has been submitted, it just didn't go anywhere -
so it feels fine to log this, and it is simpler to log in this case.