Change Step#goto_page_slug to private
The goto_page_slug is only used within the next_page_slug_after_routing
method. It's currently public. The method can also be simplified.

This commit:
- changes the name of the method to goto_condition_page_slug
- adds a parameter for the condition.
- makes the method private

Because the method is now private and only called from the next_page_slug_after_routing
method, the tests are not needed. Instead the tests for
next_page_slug_after_routing are expanded to include checking for the
CHECK_YOUR_ANSWERS_PAGE_SLUG condition.