Add default to Step#next_page_slug_after_routing
To support more complex routing in forms, form creators will be able to
add a "default" condition to a page with any question type. This will
unconditionally route them to another page when they answer the
question, regardless of their answer.

This commit allows the first condition to be set as the default. To do
so the condition must have an empty answer value.

It also ensures that pages with question types other than selection can
have routing_conditions. The code previously assumed that any steps with
routing conditions contained a question with the selection attribute.

The code change is maybe overly strict about conditions and the order they must
be in.

In order to keep this change focused on the current problem, the code
only handles the current case of a single condition for any page.

When changes to the admin are complete it might be clearer how we will
represent multiple routing conditions.

The routing conditions should probably be given their own model
and the evaluation logic might be moved to it's own class, outside of
the service.

Tackling those at this point is probably too early.