Simplify 'None of the above' logic
We're doing quite a lot of converting back and forth between symbol and
string values for the 'None of the above' options. We initially did this
because 'None of the above' is a special value for these selection
questions, but in practice it doesn't really do anything other than make
the logic a bit more complicated.

This commit removes this symbol logic so that the None of the above
value is just the string "None of the above", which is the same way we
handle other selection options.