Add id to submission reference number
When the submission type of a form is S3, we save the form to a specified bucket, using the submission reference number as part of the key.
When running end to end tests, we want to be able to submit a form, then look into an S3 bucket to find the submission.

Adding an ID here allows the end to end tests to easily find the reference number and use it to look for the submitted form in S3.