Add submissions model
Add table and model to store submission data.

The intention is to store submission data in this table when the user
submits their form. We would then have a worker job that looks up the
submission data in the database to send the submission email (or to
S3).

When using SES, we will need to keep the data after we have made the
request to send the email as it is still possible for the email to
bounce. We should then make sure we are removing data when we know the
email has been successfully sent or a period of time has elapsed and
we have not received a notification that it has bounced.