Upload file to S3 when answer is saved
Add a new `update_answer` method to question_base to set the answer
attributes on the question. This is overridden by the `file` model to
perform the upload to S3 when the answer is saved.

We do not want to save the file itself to the session, so we set this
to `nil` on the model as this model is used to set the answer data on
the session. Instead, set `original_filename` which is the name the
file was uploaded with and `uploaded_file_key` which is the S3 object
key on the session.

Make it so that we display the `original_filename` back to the user on
the "Check your answers" page.

We generate a unique key for the file in S3 using UUID to ensure there
is no chance of 2 files being uploaded with the same key.