Add SES mailer delivery method
Add a delivery method to send emails using Amazon SES and register the
delivery method. This will send an email by converting the
Mail::Message into a raw MIME email. We need to use this method as it
is the only way we are able to send the email with attachments.

This will use mailer templates configured for the mailer to format the
email content. Add an ApplicationMailer class that allows this. This
will use the default mailer layouts that already exist in the project,
which we'll later update to have GOV.UK styling.