Add SES form submission mailer
Add a mailer to send the submission email using AWS SES. The mailer
action method accepts the formatted answer content, delivery options
and any files to attach.

A view is added for the email template.

The tests are configured to use a delivery method that isn't SES so
we can appropriately test this. To accomplish this, we add a
configuration for the delivery method in `application.rb` and
`test.rb`.