Add method to determine whether form has file upload question
Add a method to return whether a form has a file upload question.
This will be used to determine whether we should send the email using
SES or Notify. Eventually we will send all emails using SES, but for
now we only want to send form submissions that include uploaded files
this way.