Add service for sending submissions using SES
Add a service to mirror `NotifySubmissionService` to send a form
submission using the SES mailer.

Add a helper class for formatting the question/answer content in the
email, mirroring `NotifyTemplateFormatter`. Haven't added tests for
this yet, as we'll want to do a little more work on this to ensure
we are formatting the HTML in a safe way.