Attach uploaded files to submission email
When sending the submission email, look for any completed steps that
are file upload questions. For these questions, retrieve the file from
S3 and pass it to the mailer to be attached to the file.

We attach files with the filename given to the file when it was
originally uploaded.

This assumes that all files are uploaded with unique filenames - which
may not be the case in reality. We will need to ensure that files have
unique names but this is not addressed in this commit.