Upload to S3 after validation is performed
We were uploaded the file to S3 for a file upload question when
`step#update!` was called. However, this was happening before the
question model was validated.

Revert `step#update!` to its orginal behaviour, which is to assign the
question model attributes from the request parameters then validate
the model.

Introduce a new `before_save` method on `QuestionBase` to perform any
actions that need to be applied before the answer is saved to the
session. Implement this for the `File` question to perform the upload
to S3 and assign the attributes on the model that we want to be saved
to the session.