Use fetch rather than require when getting question params
In the `page_controller#save` controller action, use the `fetch`
method rather than the `require` method when reading the question
params from the request. This is because the `require` method will
raise an exception if the `:question` parameter is not present.
However, for file upload questions the `:question` parameter will not
be present if no file to upload has been selected. We want to apply
the model validation and display a validation error to the user in
this case rather than raising an exception.