Allow no file to be selected when question is optional
Handle the case where a file upload question is optional and a user
has not provided a file. When no file has been selected, do not
attempt to upload the file to S3, and instead set the
`orignal_filename` attribute on the file question model to be a blank
string. We need to set a value on an attribute so when we serialize
the answer to the session, we have something to save to indicate the
question has been answered and the user can proceed to the next
question.

When sending the submission email, do not attempt to download a file
from S3 to include as an attachment in the email if a file upload
question is optional and has been skipped.