Add feature test for checking CSV submissions
The code which generates CSV files with submission details have unit
tests which stub Context with #completed_steps. These tests cover the
CsvGenerator code well but don't test how the answers to questions are
collected in the rest of the code, particularly the Context class.

Our current feature tests for filling out forms don't cover the form
submission at all.

Our policy for feature tests has been to only test values
which are externally visible. They stick to what the user would see when
using the service.

We don't check the submission email because it's hard to do without
accessing objects internal to the runner or mocking external services,
like notify.

This commit suggests using internal the internal ActionMailer class to
check the value of the CSV. This allows us to tests the contents of the
CSV being sent.

A new feature test has been added which fills out a form with an optional
question and a simple routing condition. This should give us more
confidence when making changes to the way submissions are generated.

This approach has some advantages:
- it extends the feature tests to cover an important part if the user
  journey

And some disadvantages:

- using ActionMailer to get the CSV means we may need to change the test
  code to if we use a different method to deliver submissions, for
  example AWS SES.
- the tests don't cover the full submission process. They don't cover
  the Notify API call etc.