Stub all_steps in AwsSesSubmissionService specs
The specs for AwsSesSubmissionService test submissions with and without
CSV files.

This commit stubs method all_steps in the current context so that
all_steps can be used in the CsvGenerator code without breaking these
tests.

In the specs, the values are the same because there is only one step.