Add file review component
Add a component containing the table to review the uploaded file name
along with a remove button for file upload questions. We're using a
component that inherits the Question::Base component as we need to use
the methods on this to render the question text heading and hint text,
if it's present.

Co-authored-by: Jamie Wilkinson <jamie.wilkinson1@digital.cabinet-office.gov.uk>