Add review file page
Add routes and page to review a file after it's been uploaded. This
page plays back the uploaded filename with a button to remove the
file. The continue button on the page posts to a route which will just
redirect to the next question. We do not need to save anything at this
point as the answer has already been saved.

The route to remove the file has been added, but we have not
added the implementation for this yet.

Co-authored-by: Jamie Wilkinson <jamie.wilkinson1@digital.cabinet-office.gov.uk>