Implement remove file route
Implement the remove file route, which is POSTed to when the user
clicks the remove button on the file review page. This calls S3 to
delete the file and removes the answer from the session.

This does not take into account the the question having multiple
answers, as we do not currently support "add another answer" for file
upload questions. The logic to remove the answer from the session will
need to be modified if we want to support this in the future.

Co-authored-by: Jamie Wilkinson <jamie.wilkinson1@digital.cabinet-office.gov.uk>