Use real file instead of tempfile in feature test
In the file upload feature test, we were using a TempFile. This failed
after the file type validation was added, because the TempFile returns a
 of  which trips the validation.

This commit replaces the TempFile with a real, empty text file, which is
created in the  folder and deleted after the test completes. This
should make the test closer to real use of the feature.