Show the review page for completed file upload questions
After saving the answer to a question, if the question is a file
upload question, and is not an optional question that has been
skipped, redirect to the new review file route.

We've added a new private method `save_redirect_path` for this logic,
as we use the `next_page` method in other subclasses of the
PageController, such as the ReviewFileController, and in these use
cases, we always want to go to the next step in the form.

Co-authored-by: Jamie Wilkinson <jamie.wilkinson1@digital.cabinet-office.gov.uk>