Redirect to the review file page for an answered file question
If attempting to visit the page for a file upload question, redirect
to the review file page if the question has already been answered.
On this page, the user is shown the file that has already been
uploaded with the option to remove the file and then upload another
one.