Allow selection question to return values when new
When the selection question is configured to be a checkbox and
initialised without any value for selection calling .show_answer_in_csv
will raise NoMethodError.

This happens because selection_without_blanks calls .reject on selection
without checking it's value.

This doesn't happen when a user fills out a form with a checkbox
question because params are always provided and selection is always set
to [] if no boxes are checked.

It does happen when the Question::Selection is initialised without
providing a value.

This happens in two cases:
- when a form creator is previewing a form and changes a question type
- when a CSV is generated for a form with routes and should show all
  questions in a form

The second case caused submissions to fail for a real form when changes
to the CSV generator were deployed.

This commit fixes the issue by providing a default value for selection.

The tests obscured this issue by using a before block to always set the
selection value. They've been changed to make it more explicit.