Change feature test to cover incident case
This commit changes the feature test used to check the CSV submission
data for a form.

It does this in response to an incident which caused submissions to
fail for a form with specific question types and routing.

The form which caused the incident contained a question with a checkbox
question. The changes here introduce a checkbox question into the tests
which is skipped through routing.

There are some other changes to the tests which will make it clearer
that the issue is fixed in this PR.

The email is now checked after the submission page is shown to the user.

This will allow us to change the test to expect a submission error page,
which is what form fillers saw in prod when the CSV failed to generate.