Update review file table to appear similar to a summary list
The summary list component in the design system only allows use of a
link in the action column. This is incompatible with the convention to
always use a button for something where action will be immediately
taken, as is the case for removing a file. Unfortunately, this means we
need to use a regular table that looks similar to a summary list,
instead of the summary list component itself.

If we wanted to change this, we could instead link to something akin to
a confirm file removal page, which would then use a submit button to take
action.