Move file size guidance into details component
We had some existing guidance about file size displayed directly
underneath the file upload input. It makes sense for this guidance to be
part of the 'What files can I upload?' guidance added for file types, so
this commit moves it into the same `Details` component as the file types
guidance.