Add file removal confirmation page
This adds a new page to confirm whether a file should be removed. The
motivation for this is to match the pattern we use elsewhere (add
another answer) where a user action requires a single submit button.
This also allows us to use the summary list on the review file page.

I'd like to refactor the input object to share with remove answer, as
they are functionally identical, but this will be in a subsequent PR.