Use div instead of blockquote
We were using a blockquote element here (which is commonly used for
inset text and I think the default element output by Notify for `>`
markdown input). However, semantically it isn't a quote, so now that we
have more control over the markup it makes sense to make it a `div`,
which is less semantically misleading.