Add method to poll S3 for scan status
Add a method to the FileUploadS3Service that will poll the object in
S3 until a tag with key GuardDutyMalwareScanStatus is found with the
result of the GuardDuty scan.

Initially, we will attempt polling a maximum of 20 times with a 500
millisecond wait between each attempt. When testing locally it was
seen that the tag was usually found after around 4 attempts.

The maximum number of polling attempts and the wait time have been
added as configuration settings just so we can change these in the
tests.